Update the readme file
I find the go report card badge is really useful for reminding me to
click it from time to time. So that I can see any new mistakes I've made
recently in the code.

Also, the `GO15VENDOREXPERIMENT` section is removed since it is no longe
relevant with the release of go1.7.