Temporary skip a test which failse because of a race cond
In the current implementation of the library it is impossible to wait
until all files in the library path have been scanned. The problem is
especially evident when files are added after the initial scan and
expect the FS watchers to catch them. Because of this the tests fails
some times and some - does not. For the moment the best think I could
think of is to skip it altogether. At future date a better mechanism for
scanning might make possible bringing this test back.