Use `context` based job and application control
The context lib is used liberally in places where a lot of waiting
groups and closing channels were used before. This greatly simplifies
the code and removes a plethora of different mechanism for the "stop
yourself when asked to" tasks which were previously employed.

At the moment the deadline feature of the contexts is not used. But the
groundwork is laied down so that it would be easy to add deadlines in
the HTTP handlers and library scanners.

There is one major place where the context library is not used but would
be of gread help. This is the HTTP server's graceful stop.
Unfortunately I don't see a good way to do it with a context at the
moment.