Add the "download_artwork" configuration
If set to `false` (which is the default) no artwork will be downloaded
from the internet. One must explicitly set it to `true` in order to
enable searching in the Cover Art Archive.

This default is chosen because HTTPMS does not to unwittingly start to
emit information about its users. It protects the user's privacy as best
as possible.