Move to Go modules and bid deb farewell
Dep has been an excellent dependency manager. Sadly we have to use the
modules now which will take their time to become as good. For example,
why would I need all the .travis.yml, .gitignore and other stuff in my
repo? Because the `go mod` command is half-baked, that's why!