Fix wrong association of tracks and albums
For some reason the .LastInsertId() of the sql.Result is not returning
the actual ID of the last inserted record. I found this by reading the
logs from my installation. Here is what I saw

    Inserted album id: 1182, name: Album-X, path: /path/album-X
    ...
    Inserted id: 1183, name: t7, album ID: 1183, fs_path: /path/album-X/07.mp3
    Inserted album id: 1183, name: Album-Y, path: /path/album-Y
    Inserted id: 1184, name: t1, album ID: 1183, fs_path: /path/album-Y/01.mp3
    Inserted id: 1184, name: t2, album ID: 1184, fs_path: /path/album-Y/02.mp3
    Inserted id: 1184, name: t3, album ID: 1184, fs_path: /path/album-Y/03.mp3
    Inserted id: 1184, name: t4, album ID: 1184, fs_path: /path/album-Y/04.mp3
    Inserted id: 1184, name: t5, album ID: 1184, fs_path: /path/album-Y/05.mp3
    Inserted id: 1184, name: t6, album ID: 1184, fs_path: /path/album-Y/06.mp3

Note how the first inserted track supposedly has the same ID as its
album which have been inserted right before it. At the same time the
logs say that Album-X has been inserted with ID 1182. The same goes with
the next album in the logs. LastInsertId() reports it to be inserted
with ID 1183 and all of the tracks after it report 1184.

After inspection the actual database file it turns out:

    * Album-X has ID 1183
    * Album-Y has ID 1184
    * None of the tracks' IDs reported are nowhere near correct

This all happens becuase .LastInsertId() returns seemingly random ID
from one of the last few inserts. This is most probably bug in the
go-sqlite3 library. But unfortunately I cannot reproduce it easily.

All of this lead to tracks associated with wrong albums and by similar
mechanism albums associated with wrong artists. See the logs example and
the track "t1" from Album-Y. It gets associated with the actual Album-X.

To combat this one additional SELECT is done after inserting in order to
get the real ID. The result of .LastInsertId() is checked and any
mismatch is logged. But it is disregarded otherwise.

In order to increase the likelihood that the SELECT will work as
expected additional guarantees for uniqueness have been added to the
database schema. Previously it was entirely possible for duplicate
records to be inserted.

As it stands race conditions are still possible since there is no
transaction held between checking wether something is already in the
database and then inserting it. Two different go routines might decide
to insert the same entity anyway. The synchronious database worker
does not help in this instance. Further refinement of the inserting
process will follow. Possilby relving around the new upsert feature of
SQLite which was not present at the time of writing this originally.