Add statistics for tracks and albums
These include time of last playing and play counter for tracks and
albums. For this to happen statistics are updated on every stream via
the Euterpe API and through the new "scrobble" Subsonic endpoint. Note
that in both cases Euterpe tries to prevent updating the statistics way
too fast by buggy clients.